libcusmm/kernels: fixes and improve asserts
