+++ src/acc/libsmm_acc/libcusmm/kernels/cusmm_dnt.py
@@ -9,6 +9,26 @@
 ####################################################################################################
 
 
+# ===============================================================================
+#  Computing helpers
+import numpy as np
+
+
+def round_up_to_nearest_multiple(x, step):
+    result = np.where(x % step == 0, x, x + step - x % step).astype(float)
+    if result.size == 1:
+        result = result.item()  # extract single element of numpy array
+    return result
+
+
+def round_down_to_nearest_multiple(x, step):
+    result = np.where(x % step == 0, x, x - x % step).astype(float)
+    if result.size == 1:
+        result = result.item()  # extract single element of numpy array
+    return result
+
+
+# ===============================================================================
 class Kernel:
 
     def __repr__(self):
