+++ src/acc/libsmm_acc/libcusmm/kernels/cusmm_dnt_medium.py
@@ -10,8 +10,7 @@
 
 
 import numpy as np
-from kernels.cusmm_dnt import Kernel
-from kernels.cusmm_dnt_helper import round_up_to_nearest_multiple
+from kernels.cusmm_dnt import Kernel, round_up_to_nearest_multiple
 
 
 class Kernel_dnt_medium(Kernel):
